+++ akka/src/main/scala/com/arkondata/opentracing/akka/AskTracing.scala
@@ -1,0 +1,28 @@
+package com.arkondata.opentracing.akka
+
+import akka.actor.ActorRef
+import akka.pattern
+import akka.util.Timeout
+import cats.effect.{ Async, ContextShift }
+import cats.syntax.flatMap._
+import cats.syntax.functor._
+
+import com.arkondata.opentracing.Traced
+import com.arkondata.opentracing.syntax._
+
+object AskTracing {
+  class Ops[F[_]: Async: ContextShift: Traced](ref: ActorRef, message: Any, sender: ActorRef)
+                                              (implicit timeout: Timeout) {
+    def traced: F[Any] = trace0
+
+    def trace(op: String, tags: Traced.Tag*): F[Any] = trace0.trace(op, tags: _*)
+
+    private def trace0: F[Any] =
+      for {
+        ctx <- Traced.currentSpan.context
+        res <- Async.fromFuture(
+                Async[F].delay{ pattern.ask(ref, TracedMessage(message, ctx), sender) }
+              )
+      } yield res
+  }
+}
